requirements.txt
google-api-python-client 
google-auth-oauthlib 
requests 
selenium 
python-dotenv